+++ src/test/java/wooteco/subway/service/StationServiceTest.java
@@ -17,9 +17,6 @@ import wooteco.subway.repository.SubwayRepository;
 import wooteco.subway.repository.dao.LineDao;
 import wooteco.subway.repository.dao.SectionDao;
 import wooteco.subway.repository.dao.StationDao;
-import wooteco.subway.repository.dao.jdbc.JdbcLineDao;
-import wooteco.subway.repository.dao.jdbc.JdbcSectionDao;
-import wooteco.subway.repository.dao.jdbc.JdbcStationDao;
 import wooteco.subway.repository.exception.station.DuplicateStationNameException;
 import wooteco.subway.service.dto.station.StationResponse;
 
@@ -32,9 +29,9 @@ class StationServiceTest {
 
     @BeforeEach
     void setUp() {
-        LineDao lineDao = new JdbcLineDao(dataSource);
-        SectionDao sectionDao = new JdbcSectionDao(dataSource);
-        StationDao stationDao = new JdbcStationDao(dataSource);
+        wooteco.subway.repository.dao.LineDao lineDao = new LineDao(dataSource);
+        wooteco.subway.repository.dao.SectionDao sectionDao = new SectionDao(dataSource);
+        wooteco.subway.repository.dao.StationDao stationDao = new StationDao(dataSource);
         StationRepository stationRepository = new SubwayRepository(lineDao, sectionDao, stationDao);
         this.stationService = new StationService(stationRepository);
     }
